requirements.txt
praw~=7.7.1
pyyaml~=6.0.1
pandas~=1.3.5
tenacity~=8.2.3
requests
pymongo==4.6.1
psycopg2-binary==2.9.9
apache-flink==1.18.1

prawcore~=2.4.0
pyflink~=1.0